Added RI-FILE followed by EXEC-FLAG-ARG scheme
